sys/ofed: remove all diffs
There is no way this code works and these diffs weren't close enough to
matter.  Drop them until someone is ready for make a concerted effort.

+++ sys/ofed/drivers/infiniband/core/ib_ucma.c
@@ -1726,7 +1726,7 @@ static int ucma_close(struct inode *inode, struct file *filp)
 }
 
 static long
-ucma_ioctl(struct file *filp, unsigned int cmd, uintptr_t arg)
+ucma_ioctl(struct file *filp, unsigned int cmd, unsigned long arg)
 {
 
 	switch (cmd) {
